+++ native_library/android/AdMobANE/app/src/main/java/com/tuarua/admobane/RewardedVideoController.kt
@@ -24,6 +24,7 @@ import com.google.android.gms.ads.reward.RewardItem
 import com.google.android.gms.ads.reward.RewardedVideoAd
 import com.google.android.gms.ads.reward.RewardedVideoAdListener
 import com.google.gson.Gson
+import com.tuarua.admobane.Position.*
 import com.tuarua.frekotlin.FreKotlinController
 
 class RewardedVideoController(override var context: FREContext?) : FreKotlinController, RewardedVideoAdListener {
@@ -69,11 +70,11 @@ class RewardedVideoController(override var context: FREContext?) : FreKotlinCont
     }
 
     override fun onRewardedVideoAdClosed() {
-        sendEvent(Constants.ON_CLOSED, gson.toJson(AdMobEvent(Position.REWARD)))
+        sendEvent(Constants.ON_CLOSED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdLeftApplication() {
-        sendEvent(Constants.ON_LEFT_APPLICATION, gson.toJson(AdMobEvent(Position.REWARD)))
+        sendEvent(Constants.ON_LEFT_APPLICATION, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdLoaded() {
@@ -81,30 +82,30 @@ class RewardedVideoController(override var context: FREContext?) : FreKotlinCont
         if (_showOnLoad) {
             av.show()
         }
-        sendEvent(Constants.ON_LOADED, gson.toJson(AdMobEvent(Position.REWARD)))
+        sendEvent(Constants.ON_LOADED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdOpened() {
-        sendEvent(Constants.ON_OPENED, gson.toJson(AdMobEvent(Position.REWARD)))
+        sendEvent(Constants.ON_OPENED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewarded(rewardItem: RewardItem?) {
         //trace("onRewarded")
         if (rewardItem != null) {
             sendEvent(Constants.ON_REWARDED,
-                    gson.toJson(AdMobEventWithReward(Position.REWARD, amount = rewardItem.amount, type = rewardItem.type)))
+                    gson.toJson(AdMobEventWithReward(REWARD.ordinal, amount = rewardItem.amount, type = rewardItem.type)))
         } else {
-            sendEvent(Constants.ON_REWARDED, gson.toJson(AdMobEventWithReward(Position.REWARD)))
+            sendEvent(Constants.ON_REWARDED, gson.toJson(AdMobEventWithReward(REWARD.ordinal)))
         }
 
     }
 
     override fun onRewardedVideoStarted() {
-        sendEvent(Constants.ON_VIDEO_STARTED, gson.toJson(AdMobEvent(Position.REWARD)))
+        sendEvent(Constants.ON_VIDEO_STARTED, gson.toJson(AdMobEvent(REWARD.ordinal)))
     }
 
     override fun onRewardedVideoAdFailedToLoad(p0: Int) {
-        sendEvent(Constants.ON_LOAD_FAILED, gson.toJson(AdMobEvent(Position.REWARD, p0)))
+        sendEvent(Constants.ON_LOAD_FAILED, gson.toJson(AdMobEvent(REWARD.ordinal, p0)))
     }
 
     override val TAG: String