+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.4
 - Added Reward Videos
+- Fxed Position enum on Android
 - Updated to FreSwift 1.2.0
 - Updated to FreKotlin 0.0.10